Добавляет контроллер board
